requirements.txt
aiosqlite==0.21.0
alembic==1.16.2
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
apprise==1.9.3
asgi-lifespan==2.1.0
asyncpg==0.30.0
attrs==25.3.0
backtrader==1.9.78.123
blinker==1.9.0
cachetools==6.1.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click>=8.0.0
cloudpickle==3.1.1
colorama==0.4.6
colorlog==6.9.0
coolname==2.2.0
cryptography==45.0.4
dateparser==1.2.1
docker==7.1.0
exceptiongroup==1.3.0
fastapi==0.115.13
fsspec==2025.5.1
flask==3.0.3
gitdb==4.0.12
GitPython==3.1.44
graphviz==0.21
greenlet==3.2.3
griffe==1.7.3
h11==0.16.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
humanize==4.12.3
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.7.0
iniconfig==2.1.0
Jinja2==3.1.6
jinja2-humanize-extension==0.4.0
joblib==1.5.1
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
loguru==0.7.3
Mako==1.3.10
Markdown==3.8.2
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mypy_extensions==1.1.0
narwhals==1.44.0
numpy>=1.24.0
oauthlib==3.3.1
opentelemetry-api==1.34.1
optuna==4.4.0
orjson==3.10.18
packaging==25.0
pandas>=1.5.0
pandera==0.24.0
pathspec==0.12.1
pillow==11.2.1
pluggy==1.6.0
prefect==3.4.6
prometheus_client==0.22.1
protobuf==6.31.1
pyarrow==20.0.0
pycparser==2.22
pydantic==2.11.7
pydantic-extra-types==2.10.5
pydantic-settings==2.10.0
pydantic_core==2.33.2
pydeck==0.9.1
Pygments==2.19.2
pytest>=7.0.0
python-dateutil>=2.8.2
python-dotenv>=1.0.0
python-slugify==8.0.4
python-socks==2.7.1
pytz==2025.2
pywin32==310
PyYAML>=6.0
readchar==4.2.1
referencing==0.36.2
regex==2024.11.6
requests>=2.31.0
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rich==14.0.0
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn>=1.3.0
scipy==1.16.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
streamlit>=1.28.0
ta>=0.10.0
ta-lib>=0.4.24
tenacity==9.1.2
text-unidecode==1.3
threadpoolctl==3.6.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
typeguard==4.4.4
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
tzlocal==5.3.1
ujson==5.10.0
urllib3==2.5.0
uv==0.7.14
uvicorn==0.34.3
watchdog==6.0.0
websockets==15.0.1
whenever==0.8.6
win32_setctime==1.2.0
zipp==3.23.0
psutil==5.9.0
